// ts/enums/character/attribute.enum.ts
export enum attribute {
    WEAPON_SKILL = 'weapon_skill',
    BALLISTIC_SKILL = 'ballistic_skill',
    STRENGTH = 'strength',
    THOUGHNESS = 'thoughness',
    AGILITY = 'agility',
    INTELIGENCE = 'inteligence',
    WILL_POWER = 'will_power',
    CHARISMA = 'charisma',
    WOUNDS = 'wounds',
    ATTACKS = 'attacks'
};